+++ src/Entity/PaperConflicts.php
@@ -70,7 +70,7 @@ class PaperConflicts
 
     #[ORM\ManyToOne( fetch: 'EAGER', inversedBy: 'conflicts')]
     #[ORM\JoinColumn(name: 'paper_id', referencedColumnName: 'PAPERID', nullable: true)]
-    private ?Papers $papers = null;
+    private ?Paper $papers = null;
 
     public function getCid(): ?int
     {
@@ -137,12 +137,12 @@ class PaperConflicts
         return $this;
     }
 
-    public function getPapers(): ?Papers
+    public function getPapers(): ?Paper
     {
         return $this->papers;
     }
 
-    public function setPapers(?Papers $papers): self
+    public function setPapers(?Paper $papers): self
     {
         $this->papers = $papers;